alpha 0.9.0 - finished core fuctionality, now can use multiple pdfs or img_seqs, qol

+++ src/main.cpp
@@ -1,3 +1,4 @@
+#include "cpp/poppler-document.h"
 #include "cpp/poppler-global.h"
 #include "poppler.hpp"
 #include "opencv.hpp"
@@ -9,19 +10,20 @@
 #include <opencv2/imgcodecs.hpp>
 
 using std::string;
-// using std::vector;
 namespace fs = std::filesystem;
 
 int main(int argc, char **argv) {
+    bool keep = false;
+    bool is_pdf = false;
+    bool is_seq = false;
     int width = 1280;
     int height = 720;
     double fps = 1.0;
     double spp = 1.0;
-    string vid_fmt = "MP4";
     Style style = Style::SEQUENCE;
-    string pdf_path = "";
-    string img_seq_dir = "";
-    bool keep = false;
+    string vid_fmt = "MP4";
+    vector<string> pdf_paths = {};
+    vector<string> seq_dirs = {};
 
     if (argc < 2) {
         std::cerr << HELP_TXT << std::endl;
@@ -35,32 +37,36 @@ int main(int argc, char **argv) {
             std::cout << HELP_TXT << std::endl;
             return 0;
         } else if ((int)arg.find(".pdf") > -1) {
-            if (img_seq_dir != "") {
+            if (is_seq) {
                 std::cerr << "<!> Error: Cannot convert Image Sequence and PDF at the same time." << std::endl;
                 return 1;
+            } else {
+                is_pdf = true;
             }
             if (!fs::exists(arg)) {
-                std::cerr << "<!> Error: PDF path does not exist." << std::endl;
+                std::cerr << "<!> Error: '" << arg << "' does not exist." << std::endl;
                 return 1;
             }
             if (arg.substr(arg.length() - 4, 4) != ".pdf") {
-                std::cout << "<!> Error: PDF path not a .pdf file" << std::endl;
+                std::cout << "<!> Error: '" << arg << "' is not a PDF file." << std::endl;
                 return 1;
             }
-            pdf_path = arg;
+            pdf_paths.push_back(arg);
         } else if ((int)arg.find('/') > -1) {
-            if (pdf_path != "") {
+            if (is_pdf) {
                 std::cerr << "<!> Error: Cannot convert PDF and Image Sequence at the same time." << std::endl;
                 return 1;
+            } else {
+                is_seq = true;
             }
             if (!fs::is_directory(arg)) {
-                std::cerr << "<!> Error: directory path dosen't exist. Try again." << std::endl;
+                std::cerr << "<!> Error: '" << arg <<  "' is not a valid directory." << std::endl;
                 return 1;
             }
             if (arg[arg.length() - 1] != '/') {
                 arg.push_back('/');
             }
-            img_seq_dir = arg;
+            seq_dirs.push_back(arg);
         } else if (arg == "-r") {
             i++;
             width = std::stoi(argv[i]);
@@ -89,15 +95,26 @@ int main(int argc, char **argv) {
         }
     }
 
-    // pdf_loading_test();
-    // return 1;
-
     // Info presented before continuing.
-    if (pdf_path != "") {
-        std::cout << "PDF Path: " << pdf_path << std::endl;
+    if (pdf_paths.size() > 0) {
+        std::cout << "PDF Paths (" << pdf_paths.size() << "): ";
+        for (size_t i = 0; i < pdf_paths.size(); i++) {
+            if (i > 0) {
+                std::cout << " + ";
+            }
+            std::cout << "" + pdf_paths[i];
+        }
+        std::cout << std::endl;
     }
-    if (img_seq_dir != "") {
-        std::cout << "Image Sequence Dir: " << img_seq_dir << std::endl;
+    if (seq_dirs.size() > 0) {
+        std::cout << "Sequence Directories (" << seq_dirs.size() << "): ";
+        for (size_t i = 0; i < seq_dirs.size(); i++) {
+            if (i > 0) {
+                std::cout << " + ";
+            }
+            std::cout << "" + seq_dirs[i];
+        }
+        std::cout << std::endl;
     }
     std::cout << "Width: " << width << "\nHeight: " << height << "" << std::endl;
     std::cout << "FPS: " << fps << std::endl;
@@ -134,9 +151,12 @@ int main(int argc, char **argv) {
 
     // Application Logic
     // converting pdf to video
-    if (pdf_path != "") {
-        poppler::document *pdf = poppler::document::load_from_file(pdf_path);
-        string frames_dir = get_frames_dir(pdf_path);
+    if (is_pdf) {
+        poppler::document *pdf = poppler::document::load_from_file(pdf_paths[0]);
+        std::vector<cv::Mat> images = {};
+        string frames_dir = get_frames_dir(pdf_paths[0]);
+        string fmt_path = format_path(pdf_paths[0]);
+        string output = fmt_path.substr(0, fmt_path.length() - 4) + "." + vp.fmt;
 
         // sets viewport's resolution to full res of first pdf page
         poppler::rectf rect = pdf->create_page(0)->page_rect(poppler::media_box);
@@ -148,47 +168,7 @@ int main(int argc, char **argv) {
         }
 
         std::cout << "Loading pdf pages..." << std::endl;
-        std::vector<cv::Mat> images = get_images(pdf, style, vp);
-
-        make_frames_dir(frames_dir);
-
-        if (style == Style::SCROLL) {
-            cv::Mat long_img = get_long_image(images, vp);
-            std::cout << "Generating video frames..." << std::endl;
-            generate_scroll_frames(frames_dir, pdf->pages(), long_img, vp);
-        } else if (style == Style::SEQUENCE) {
-            std::cout << "Generating video frames..." << std::endl;
-            generate_sequence_frames(frames_dir, images, vp);
-        }
-
-        string fmt_path = format_path(pdf_path);
-        string output = fmt_path.substr(0, fmt_path.length() - 4) + "." + vp.fmt;
-        generate_video(frames_dir, output, vp);
-
-        if (!keep) {
-            delete_dir(frames_dir);
-        }
-
-        std::cout << "Finished!" << std::endl;
-    }
-
-    // converting image sequence to video
-    if (img_seq_dir != "") {
-        string frames_dir = get_frames_dir(img_seq_dir);
-        std::map<int, string> img_map = get_image_sequence_map(img_seq_dir);
-
-        // sets viewport res to resolution of img
-        cv::Mat img = cv::imread(img_map[-1]);
-        if (width == -1) {
-            vp.width = img.cols % 2 == 0 ? img.cols : img.cols + 1;
-        }
-        if (height == -1) {
-            vp.height = img.rows % 2 == 0 ? img.rows : img.rows + 1;
-        }
-        img_map.erase(-1);
-
-        std::cout << "Loading images..." << std::endl;
-        vector<cv::Mat> images = get_images(img_map, style, vp);
+        images = get_images(pdf_paths, style, vp);
 
         make_frames_dir(frames_dir);
 
@@ -201,14 +181,45 @@ int main(int argc, char **argv) {
             generate_sequence_frames(frames_dir, images, vp);
         }
 
-        string fmt_dir = format_path(img_seq_dir);
-        string output = fmt_dir.substr(0, fmt_dir.length() - 1) + "." + vp.fmt;
         generate_video(frames_dir, output, vp);
+        if (!keep) delete_dir(frames_dir);
+        std::cout << "Finished!" << std::endl;
+    }
 
-        if (!keep) {
-            delete_dir(frames_dir);
+    // converting image sequence to video
+    if (is_seq) {
+        std::map<int, string> img_map = get_image_sequence_map(seq_dirs);
+        vector<cv::Mat> images = {};
+        string frames_dir = get_frames_dir(seq_dirs[0]);
+        string fmt_dir = format_path(seq_dirs[0]);
+        string output = fmt_dir.substr(0, fmt_dir.length() - 1) + "." + vp.fmt;
+
+        // sets viewport res to resolution of img
+        cv::Mat img = cv::imread(img_map[-1]);
+        if (width == -1) {
+            vp.width = img.cols % 2 == 0 ? img.cols : img.cols + 1;
+        }
+        if (height == -1) {
+            vp.height = img.rows % 2 == 0 ? img.rows : img.rows + 1;
+        }
+        img_map.erase(-1);
+
+        std::cout << "Loading images..." << std::endl;
+        images = get_images(img_map, style, vp);
+
+        make_frames_dir(frames_dir);
+
+        if (style == Style::SCROLL) {
+            cv::Mat long_img = get_long_image(images, vp);
+            std::cout << "Generating video frames..." << std::endl;
+            generate_scroll_frames(frames_dir, images.size(), long_img, vp);
+        } else if (style == Style::SEQUENCE) {
+            std::cout << "Generating video frames..." << std::endl;
+            generate_sequence_frames(frames_dir, images, vp);
         }
 
+        generate_video(frames_dir, output, vp);
+        if (!keep) delete_dir(frames_dir);
         std::cout << "Finished!" << std::endl;
     }
